data structures,greedy,*1100

+++ Codeforces Round #805 (Div. 3)/C. Train and Queries.cpp
@@ -73,9 +73,9 @@ void solve(void){
     ll n; cin >> n;
     vll v(n); rpt(i, 0, n) cin >> v[i];
 
-    ll sum = accumulate(all(v), 0);
+    ll sum = accumulate(all(v), (ll)0);
     if(!(sum % 2)) kill("0");
-    
+
     ll ans = inf;
     rpt(i, 0, n) 
     	ans = min(ans, setbit(v[i], v[i] % 2));
